% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcdc_browse.R
\name{bcdc_browse}
\alias{bcdc_browse}
\title{Load the B.C. Data Catalogue URL into an HTML browser}
\usage{
bcdc_browse(url = "https://catalogue.data.gov.bc.ca",
  browser = getOption("browser"), encodeIfNeeded = FALSE)
}
\arguments{
\item{url}{a non-empty character string giving the URL to be loaded. Default \code{https://catalogue.data.gov.bc.ca}.}

\item{browser}{a non-empty character string giving the name of the
    program to be used as the HTML browser.  It should be in the PATH,
    or a full path specified.  Alternatively, an \R function to be
    called to invoke the browser.

    Under Windows \code{NULL} is also allowed (and is the default), and
    implies that the file association mechanism will be used.
  }

\item{encodeIfNeeded}{Should the URL be encoded by
    \code{\link{URLencode}} before passing to the browser?  This is not
    needed (and might be harmful) if the \code{browser} program/function
    itself does encoding, and can be harmful for \samp{file://} URLs on some
    systems and for \samp{http://} URLs passed to some CGI applications.
    Fortunately, most URLs do not need encoding.}
}
\value{
A browser is opened with the B.C. Data Catalogue URL loaded: nothing is returned to the R interpreter
}
\description{
This is a wrapper around utils::browseURL with the URL for the B.C. Data Catalogue as
the default
}
\examples{
bcdc_browse()
}
\seealso{
\code{\link[utils]{browseURL}}
}
