% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-classes.R
\name{bcdc_cache_path}
\alias{bcdc_cache_path}
\title{Retrieve Default Cache Path}
\usage{
bcdc_cache_path()
}
\description{
Retrieves the default path used to cache the result of web requests. Makes
use of the \code{rappdirs} package to use cache folders
defined by each operating system
}
