% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-classes.R
\name{bcdc_cache_timeout}
\alias{bcdc_cache_timeout}
\title{Retrieve Default Cache timeout}
\usage{
bcdc_cache_timeout()
}
\description{
Retrieves the length of time that a cache of \code{\link[=collect]{collect()}}ed
web resources is kept. Default is 1 hour (3600 secons).
}
