% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcdc-web-services.R
\name{bcdc_cql_string}
\alias{bcdc_cql_string}
\title{Create CQL filter strings from sf objects}
\usage{
bcdc_cql_string(x, geometry_predicates)
}
\arguments{
\item{x}{object of class sf, sfc or sfg}

\item{geometry_predicates}{Geometry predicates that allow for spatial filtering.
bcbdc_cql_string accepts the following geometric predicates: EQUALS,
DISJOINT, INTERSECTS, TOUCHES, CROSSES,  WITHIN, CONTAINS, OVERLAPS, RELATE,
DWITHIN, BEYOND.}
}
\description{
Convenience wrapper to convert sf objects and geometric operations into CQL
filter strings which can then be supplied to the query argument in
\code{bcdc_get_geodata}. The sf object is automatically converted in a
bounding box to reduce the complexity of the wfs call. Subsequent in-memory
filtering may be need to achieve exact results.
}
\examples{
airports <- bcdc_get_geodata("bc-airports")
bcdc_cql_string(airports, "DWITHIN")

}
