% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcdc_search.R
\name{bcdc_facets}
\alias{bcdc_facets}
\title{Get the valid values for a facet (that you can use in \code{\link[=bcdc_search]{bcdc_search()}})}
\usage{
bcdc_facets(facet = c("license_id", "download_audience", "type",
  "res_format", "sector", "organization"))
}
\arguments{
\item{facet}{the facet for which to retrieve valid values}
}
\value{
a data frame of values for the selected facet
}
\description{
Get the valid values for a facet (that you can use in \code{\link[=bcdc_search]{bcdc_search()}})
}
\examples{
bcdc_facets("type")
}
