% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-classes.R
\name{bcdc_forget}
\alias{bcdc_forget}
\title{Forget (clear) the cache of objects returned by
\code{\link[=collect]{collect()}}}
\usage{
bcdc_forget()
}
\value{
\code{TRUE} if the cache existed previously and was
successfully cleared, otherwise \code{FALSE}.
}
\description{
Forget (clear) the cache of objects returned by
\code{\link[=collect]{collect()}}
}
