% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{bcdc_get_data}
\alias{bcdc_get_data}
\title{downlaod and read a dataset from a Data Catalogue resource}
\usage{
bcdc_get_data(x)
}
\arguments{
\item{x}{either a \code{bcdc_record} object (from the result of \code{bcdc_get_record()})
or a character string denoting the id of a resource (or the url).}
}
\value{
an object of a type relevant to the resource (usually a tibble or an sf object)
}
\description{
downlaod and read a dataset from a Data Catalogue resource
}
\examples{
bcdc_get_data(paste0("https://catalogue.data.gov.bc.ca/dataset/",
                     "british-columbia-greenhouse-gas-emissions/resource/",
                     "11b1da01-fabc-406c-8b13-91e87f126dec"))
bcdc_get_data("11b1da01-fabc-406c-8b13-91e87f126dec")
}
