% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcdc_browse.R
\name{bcdc_get_data_citation}
\alias{bcdc_get_data_citation}
\title{Generate a bibentry from a Data Catalogue Record}
\usage{
bcdc_get_data_citation(id)
}
\arguments{
\item{id}{the human-readable name, permalink ID, or
URL of the record.

It is advised to use the permanent ID for a record rather than the
human-readable name to guard against future name changes of the record.
If you use the human-readable name a warning will be issued once per
session. You can silence these warnings altogether by setting an option:
\code{options("silence_named_get_record_warning" = TRUE)} - which you can put
in your .Rprofile file so the option persists across sessions.}
}
\description{
Generate a bibentry object for use .bib file.
for more details
}
\examples{

try(
cite <- bcdc_get_data_citation("76b1b7a3-2112-4444-857a-afccf7b20da8")

print(cite, stype = "citation")
)
}
\seealso{
\link{utils::bibentry}
}
