% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcdc-web-services.R
\name{bcdc_get_geodata}
\alias{bcdc_get_geodata}
\title{Get data from the BC Web Feature Service}
\usage{
bcdc_get_geodata(feature_name = NULL, ...)
}
\arguments{
\item{feature_name}{Name of the feature}

\item{...}{passed to \code{sf::read_sf}}
}
\description{
Pulls features off the web
}
\examples{

bcdc_wfs("WHSE_WATER_MANAGEMENT.WRIS_DAMS_PUBLIC_SVW")
bcdc_wfs("WHSE_ENVIRONMENTAL_MONITORING.ENVCAN_HYDROMETRIC_STN_SP")
}
