% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcdc_search.R
\name{bcdc_group}
\alias{bcdc_group}
\title{Retrieve all metadata of all data within a group}
\usage{
bcdc_group(group)
}
\arguments{
\item{group}{Name of the group}
}
\description{
Returns a tibble of records by providing the group name. Groups can be view
here: https://catalogue.data.gov.bc.ca/group or accessed directly from R using
\href{`bcdc_list_groups`}{bcdc_list_groups}
}
\examples{
\donttest{
bcdc_group('environmental-reporting-bc')
}
}
