% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcdc_search.R
\name{bcdc_list}
\alias{bcdc_list}
\title{Return a full list of the names of B.C. Data Catalogue records}
\usage{
bcdc_list()
}
\value{
A character vector of the names of B.C. Data Catalogue records
}
\description{
Return a full list of the names of B.C. Data Catalogue records
}
