% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcdc_search.R
\name{bcdc_list_organizations}
\alias{bcdc_list_organizations}
\alias{bcdc_list_organization_records}
\title{Retrieve organization information for B.C. Data Catalogue}
\usage{
bcdc_list_organizations()

bcdc_list_organization_records(organization)
}
\arguments{
\item{organization}{Name of the organization}
}
\description{
Returns a tibble of organizations or records. Organizations can be viewed here:
https://catalogue.data.gov.bc.ca/organizations or accessed directly from R using \code{bcdc_list_organizations}
}
\section{Functions}{
\itemize{
\item \code{bcdc_list_organizations()}: 

}}
\examples{
\donttest{
try(
  bcdc_list_organization_records('bc-stats')
)
}
}
