% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcdc-web-services.R
\name{bcdc_preview}
\alias{bcdc_preview}
\title{Get map from the B.C. Web Mapping Service}
\usage{
bcdc_preview(x = NULL)
}
\arguments{
\item{x}{either a \code{bcdc_record} object (from the result of \code{bcdc_get_record()})
or a character string denoting the id of a resource (or the url).}
}
\description{
Get map from the B.C. Web Mapping Service
}
\examples{
\dontrun{
bcdc_preview("regional-districts-legally-defined-administrative-areas-of-bc")
bcdc_preview("points-of-well-diversion-applications")
}
}
