% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcdc-web-services.R
\name{bcdc_query_geodata}
\alias{bcdc_query_geodata}
\title{Query data from the BC Web Feature Service}
\usage{
bcdc_query_geodata(x = NULL, crs = 3005)
}
\arguments{
\item{x}{either a \code{bcdc_record} object (from the result of \code{bcdc_get_record()})
or a character string denoting the id of a resource (or the url).}

\item{crs}{the epsg code for the coordinate reference system. Defaults to \code{3005}
(B.C. Albers). See https://epsgi.io.}
}
\value{
an \code{sf} object
}
\description{
Queries features from the BC Web Feature Service. The data must be available as
a wms/wfs service. See \code{bcdc_get_record(x)$resources}).
}
\examples{

\dontrun{
bcdc_query_geodata("bc-airports", crs = 3857)
bcdc_query_geodata("bc-airports", crs = 3857) \%>\%
  filter(PHYSICAL_ADDRESS == 'Victoria, BC') \%>\%
  collect()
bcdc_query_geodata("ground-water-wells") \%>\%
  filter(OBSERVATION_WELL_NUMBER == 108)

## A moderately large layer
bcdc_query_geodata("bc-environmental-monitoring-locations")
bcdc_query_geodata("bc-environmental-monitoring-locations") \%>\%
  filter(PERMIT_RELATIONSHIP == "DISCHARGE")


## A very large layer
bcdc_query_geodata("terrestrial-protected-areas-representation-by-biogeoclimatic-unit")
}

}
