% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{bcdc_read_functions}
\alias{bcdc_read_functions}
\title{Formats supported and loading functions}
\usage{
bcdc_read_functions()
}
\description{
Provides a tibble of formats supported by bcdata and the associated function that
reads that data into R. This function is meant as a resource to determine which parameters
can be passed through the \code{bcdc_get_data} function to the reading function. This is
particularly important to know if the data requires using arguments from the read in function.
}
