% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcdc_search.R
\name{bcdc_tidy_resources}
\alias{bcdc_tidy_resources}
\title{Provide a data frame containing the metadata for all resources from a single B.C. Data Catalogue record}
\usage{
bcdc_tidy_resources(record)
}
\arguments{
\item{record}{the human-readable name, permalink ID, or
URL of the record, or a \code{bcdc_record} object (from the result of \code{bcdc_get_record()})}
}
\value{
A data frame containing the metadata for all the resources for a record
}
\description{
Provide a data frame containing the metadata for all resources from a single B.C. Data Catalogue record
}
\examples{
\donttest{
airports <- bcdc_get_record("bc-airports")
bcdc_tidy_resources(airports)
}
}
