% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcdc-web-services.R
\name{bcdc_wms}
\alias{bcdc_wms}
\title{Get map from the BC Web Mapping Service}
\usage{
bcdc_wms(feature_name = NULL, bbox = NULL)
}
\arguments{
\item{feature_name}{Name of the feature}

\item{bbox}{a string of bounding box coordinates}
}
\description{
Pulls map off the web
}
\examples{
\dontrun{
## So far only works with this layer
bbox_coords = "1069159.051186301,1050414.7675306,1074045.5446851396,1054614.0978811644"
bcdc_wms("WHSE_FOREST_VEGETATION.VEG_COMP_LYR_R1_POLY", bbox = bbox_coords)
}
}
