% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-classes.R
\name{collect.bcdc_promise}
\alias{collect.bcdc_promise}
\title{Force collection of WFS request from BC Data Catalogue}
\usage{
\method{collect}{bcdc_promise}(.data)
}
\description{
Force collection of WFS request from BC Data Catalogue
}
