% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cql-geom-predicates.R
\name{cql_geom_predicates}
\alias{cql_geom_predicates}
\alias{EQUALS}
\alias{DISJOINT}
\alias{INTERSECTS}
\alias{TOUCHES}
\alias{CROSSES}
\alias{WITHIN}
\alias{CONTAINS}
\alias{OVERLAPS}
\alias{RELATE}
\alias{DWITHIN}
\alias{BEYOND}
\title{CQL Geometry Predicates}
\usage{
EQUALS(geom)

DISJOINT(geom)

INTERSECTS(geom)

TOUCHES(geom)

CROSSES(geom)

WITHIN(geom)

CONTAINS(geom)

OVERLAPS(geom)

RELATE(geom)

DWITHIN(geom)

BEYOND(geom)
}
\arguments{
\item{geom}{an sf object}
}
\value{
a CQL expression using the bounding box of the geom
}
\description{
Functions to construct a CQL expression to be used
to filter results from \link{bcdc_get_geo_data}
}
