% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportaciones.R
\name{get_exportaciones}
\alias{get_exportaciones}
\title{Descarga las exportaciones dominicanas}
\usage{
get_exportaciones()
}
\value{
Un tibble con la serie de las exportaciones
}
\description{
Esta funcion descarga los archivos con las exportaciones
dominicanas y los combina para terminar con una serie actualizada
y desagregada por producto, tipo de producto y regimen de exportación.
El valor de las exportaciones está en millones de RD$.
}
