% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportaciones.R
\name{get_impotaciones}
\alias{get_impotaciones}
\title{Descarga las importaciones dominicanas}
\usage{
get_exportaciones()
}
\value{
Un tibble con la serie de las importaciones
}
\description{
Esta funcion descarga los archivos con las importaciones
dominicanas y los combina para terminar con una serie actualizada
y desagregada por producto, tipo de producto y regimen de importacion
}
