% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipc.R
\name{get_ipc_data}
\alias{get_ipc_data}
\title{Descarga las series del IPC}
\usage{
get_ipc_data(desagregacion = "general")
}
\arguments{
\item{desagregacion}{string indicando la desagregacion deseada. opciones:
"general", "grupos", "regiones", "subyacente", "tnt" (transable y no transable),
"articulos"}
}
\value{
Un tibble con las series del ipc con la desagregacion deseada
}
\description{
Descarga las series del IPC de la Republica Dominicana
con diferentes desagregaciones
}
\examples{
get_ipc_data()
get_ipc_data(desagregacion = "grupos")
get_ipc_data(desagregacion = "regiones")
}
