% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pib_sectores.R
\name{get_pib_sectores}
\alias{get_pib_sectores}
\title{Descarga la series del PIB por sectores de origen}
\usage{
get_pib_sectores(modalidad = "real", acumulado = FALSE, homogenea_91 = FALSE)
}
\arguments{
\item{modalidad}{String indicando la modalidad deseada
de los datos, si precios corrientes o \\u00edndicesndices de volumen
encadenados. Opciones: "nominal" y "real"}

\item{acumulado}{Logical indicando si se desea serie trimestral
o trimestral acumulada al año.}

\item{homogenea_91}{Logical indicando si se desea la serie homogenea
desde el año 1991, con base 2007.}
}
\value{
Un tibble con la serie del PIB desagregada por sectores de origen.
}
\description{
Descarga las series del PIB por sectores de origen,
con diferentes alternativas según las metodolog\\u00edas
y opciones de presentaci\\u00f3n de resultados.
}
