% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tc.R
\name{get_tcambio}
\alias{get_tcambio}
\title{Descarga series del tipo de cambio}
\usage{
get_tcambio(frecuencia = "diaria")
}
\arguments{
\item{frecuencia}{La frecuencia de la deseada de la serie: "diaria", "mensual",
"trimestral" y "anual"}
}
\value{
un tibble con las serie del tipo de cambio de comprar y venta
}
\description{
Descarga las series de tipo promedio de las operaciones de compra y venta
de d\\u00f3lar en las entidades de intermediaci\\u00f3n financiera
}
\examples{
get_tcambio(frecuencia = "mensual")
get_tcambio(frecuencia = "diaria")
}
