% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_url}
\alias{make_url}
\title{Combine url components without having to worry
about slashes}
\usage{
make_url(..., trailing_slash = FALSE)
}
\arguments{
\item{...}{url components}

\item{trailing_slash}{}
}
\value{
complete url
}
\description{
Combine url components without having to worry
about slashes
}
