% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-classes.R
\name{mutate.bcdc_promise}
\alias{mutate.bcdc_promise}
\title{Mutate columns from Web Service call}
\usage{
\method{mutate}{bcdc_promise}(.data, ...)
}
\arguments{
\item{.data}{object of class \code{bcdc_promise} (likely passed from \code{\link[=bcdc_query_geodata]{bcdc_query_geodata()}})}

\item{...}{One or more unquoted expressions separated by commas. See details.}
}
\description{
The CQL syntax to generate WFS calls does not current allow arithmetic operations. Therefore
this function exists to
}
\examples{
\dontrun{

## Mutate columns
bcdc_query_geodata("bc-airports") \%>\%
  mutate(LATITUDE * 100)
}

}
