% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-classes.R
\name{select.bcdc_promise}
\alias{select.bcdc_promise}
\title{select columns from wfs call}
\usage{
\method{select}{bcdc_promise}(.data, ...)
}
\arguments{
\item{.data}{passed from bcdc_get_geodata}

\item{...}{One or more unquoted expressions separated by commas. See details.}
}
\description{
Similar to a \code{dplyr::select} call, this allows you to select which columns you want the wfs to return.
A key difference between \code{dplyr::select} and \code{bcdata::select} is the presence of "sticky" columns that are
returned regardless of what columns are selected. If any of these "sticky" columns are selected
only "sticky" columns are returns. \code{bcdc_describe_feature} is one way to tell if columns are stick in advance
of issuing the wfs call.
}
\examples{
\dontrun{
feature_spec <- bcdc_describe_feature("bc-airports")
## Columns that can selected:
feature_spec[feature_spec$nillable == TRUE,]

## Select columns
bcdc_get_geodata("bc-airports") \%>\%
  select(DESCRIPTION, PHYSICAL_ADDRESS)

## Select "sticky" columns
bcdc_get_geodata("bc-airports") \%>\%
  select(LOCALITY)
}

}
