% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cql-translator.R
\docType{class}
\name{wfsConnection-class}
\alias{wfsConnection-class}
\alias{dbQuoteIdentifier,wfsConnection,ANY-method}
\alias{dbQuoteString,wfsConnection,ANY-method}
\title{wfsConnection class}
\usage{
\S4method{dbQuoteIdentifier}{wfsConnection,ANY}(conn, x)

\S4method{dbQuoteString}{wfsConnection,ANY}(conn, x)
}
\description{
wfsConnection class
}
\keyword{internal}
