% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcdata-package.R
\docType{package}
\name{bcdata-package}
\alias{bcdata}
\alias{bcdata-package}
\title{bcdata: Search and Retrieve Data from the BC Data Catalogue}
\description{
Search and retrieve data from the BC data catalogue
    (<https://catalogue.data.gov.bc.ca/>) directly from R. Spatial and
    non-spatial data are retrieved differently depending on the
    characteristics of the data object.
}
\author{
\strong{Maintainer}: Andy Teucher \email{andy.teucher@gov.bc.ca}

Authors:
\itemize{
  \item Sam Albers \email{sam.albers@gov.bc.ca} [contributor]
  \item Stephanie Hazlitt \email{stephanie.hazlitt@gov.bc.ca} [contributor]
}

Other contributors:
\itemize{
  \item Province of British Columbia [copyright holder]
}

}
\keyword{internal}
