% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcdata-package.R
\docType{package}
\name{bcdata-package}
\alias{bcdata}
\alias{bcdata-package}
\title{bcdata: Search and Retrieve Data from the BC Data Catalogue}
\description{
\if{html}{\figure{logo.png}{options: align='right'}}

Search, query, and download tabular and geospatial data from the 
             British Columbia Data Catalogue(<https://catalogue.data.gov.bc.ca/>). 
             Search catalogue data records based on keywords, data licence, 
             sector, data format, and B.C. government organization. View 
             metadata directly in R, download many data formats, and query 
             geospatial data available via the B.C. government Web Feature 
             Service (WFS) using dplyr syntax.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/bcgov/bcdata}
  \item Report bugs at \url{https://github.com/bcgov/bcdata/issues}
}

}
\author{
\strong{Maintainer}: Andy Teucher \email{andy.teucher@gov.bc.ca}

Authors:
\itemize{
  \item Sam Albers \email{sam.albers@gov.bc.ca} [contributor]
  \item Stephanie Hazlitt \email{stephanie.hazlitt@gov.bc.ca} [contributor]
}

Other contributors:
\itemize{
  \item Province of British Columbia [copyright holder]
}

}
\keyword{internal}
