% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe-feature.R
\name{bcdc_describe_feature}
\alias{bcdc_describe_feature}
\title{Describe a WFS feature}
\usage{
bcdc_describe_feature(x = NULL)
}
\arguments{
\item{x}{either a \code{bcdc_record} object (from the result of \code{bcdc_get_record()})
or a character string denoting the id of a resource (or the url).}
}
\description{
Describe the columns from a WFS feature. The column name, whether a column can be
separated from the record in WFS (nillable) and the tpye of column are returned.
This can be a helpful tool to examine a layer before issuing a query with \code{bcdc_get_geodata}
}
\examples{
 bcdc_describe_feature("bc-airports")
}
