% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{bcdc_get_data}
\alias{bcdc_get_data}
\title{Download and read a resource from a B.C. Data Catalogue record}
\usage{
bcdc_get_data(record, resource = NULL, ...)
}
\arguments{
\item{record}{either a \code{bcdc_record} object (from the result of \code{bcdc_get_record()})
or a character string denoting the id of a resource (or the url).

It is advised to use the permament id for a record rather than the
human-readable name to guard against future name changes of the record.
If you use the human-readable name a warning will be issued once per
session. You can silence these warnings altogether by setting an option:
\code{options("silence_named_get_data_warning" = TRUE)} - which you can set
in your .Rprofile file so the option persists across sessions.}

\item{resource}{option argument used when there are multiple data files of the same format
within the same record. See examples.}

\item{...}{arguments passed to other functions. For spatial (WMS/WFS) data these are passed to
\code{bcdc_query_geodata()}. Non spatial data is passed to a function to handle the import based
on the file extension.}
}
\value{
an object of a type relevant to the resource (usually a tibble or an sf object)
}
\description{
Download and read a resource from a B.C. Data Catalogue record
}
\examples{
\dontrun{
bcdc_get_data("bc-airports")
bcdc_get_data("bc-winery-locations")
bcdc_get_data("local-government-population-and-household-projections-2018-2027",
sheet = "Population", skip = 1)

}
}
