% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{bcdc_get_data}
\alias{bcdc_get_data}
\title{downlaod and read a dataset from a Data Catalogue resource}
\usage{
bcdc_get_data(x, format = NULL, ...)
}
\arguments{
\item{x}{either a \code{bcdc_record} object (from the result of \code{bcdc_get_record()})
or a character string denoting the id of a resource (or the url).}

\item{format}{Defaults to NULL which will first check to see if the record is a wms/wfs record.
If so an sf object is returned. Otherwise a format needs to be specified. \code{bcdc_get_record} can
be used to identify which formats are available.}

\item{...}{arguments passed to other functions.}
}
\value{
an object of a type relevant to the resource (usually a tibble or an sf object)
}
\description{
downlaod and read a dataset from a Data Catalogue resource
}
\examples{
\dontrun{
bcdc_get_data("bc-airports")
bcdc_get_data("bc-winery-locations", format = "csv")
bcdc_get_data("local-government-population-and-household-projections-2018-2027",
format = "xlsx", sheet = "Population", skip = 1)

}
}
