% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcdc-web-services.R
\name{bcdc_get_geodata}
\alias{bcdc_get_geodata}
\title{Get data from the BC Web Feature Service}
\usage{
bcdc_get_geodata(x = NULL, ..., crs = 3005)
}
\arguments{
\item{x}{either a \code{bcdc_record} object (from the result of \code{bcdc_get_record()})
or a character string denoting the id of a resource (or the url).}

\item{...}{Logical predicates with which to filter the results. Multiple
conditions are combined with \code{&}. Only rows where the condition evalueates to
\code{TRUE} are kept.}

\item{crs}{the epsg code for the coordinate reference system. Defaults to \code{3005}
(B.C. Albers). See https://epsgi.io.}
}
\value{
an \code{sf} object
}
\description{
Pulls features off the web. The data must be available as a wms/wfs service.
See \code{bcdc_get_record(x)$resources}). If the record is greater than 10000 rows,
the response will be paginated. If you are querying layers of this size, expect
that the request will take quite a while.
}
\examples{

bcdc_get_geodata("bc-airports", crs = 3857)
bcdc_get_geodata("bc-airports", PHYSICAL_ADDRESS == 'Victoria, BC', crs = 3857)
bcdc_get_geodata("ground-water-wells", OBSERVATION_WELL_NUMBER == 108)

## A very large layer
\dontrun{
bcdc_get_geodata("terrestrial-protected-areas-representation-by-biogeoclimatic-unit")
}

}
