% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcdc-web-services.R
\name{bcdc_get_geodata}
\alias{bcdc_get_geodata}
\title{Get data from the B.C. Web Feature Service (DEPRECATED)}
\usage{
bcdc_get_geodata(record = NULL, crs = 3005)
}
\arguments{
\item{record}{either a \code{bcdc_record} object (from the result of \code{bcdc_get_record()})
or a character string denoting the ID of a resource (or the url).

It is advised to use the permament ID for a record rather than the
human-readable name to guard against future name changes of the record.
If you use the human-readable name a warning will be issued once per
session. You can silence these warnings altogether by setting an option:
\code{options("silence_named_get_data_warning" = TRUE)} - which you can set
in your .Rprofile file so the option persists across sessions.}

\item{crs}{the epsg code for the coordinate reference system. Defaults to \code{3005}
(B.C. Albers). See https://epsgi.io.}
}
\value{
an \code{sf} object
}
\description{
Get data from the B.C. Web Feature Service (DEPRECATED)
}
