% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcdc-web-services.R
\name{bcdc_get_geodata}
\alias{bcdc_get_geodata}
\title{Get data from the BC Web Feature Service}
\usage{
bcdc_get_geodata(id = NULL, query = NULL, crs = 3005, ...)
}
\arguments{
\item{id}{the name of the record}

\item{query}{A valid \href{https://docs.geoserver.org/stable/en/user/tutorials/cql/cql_tutorial.html}{CQL or ECQL query}
to filter the results. Default \code{NULL} (return all objects)}

\item{crs}{the epsg code for the coordinate reference system. Default \code{3005}
(B.C. Albers). See https://epsgi.io.}

\item{...}{passed to \code{sf::read_sf}}
}
\value{
an \code{sf} object
}
\description{
Pulls features off the web. The data must be available as a wms/wfs service.
See \code{bcdc_get_record(id)$resources})
}
\examples{

bcdc_get_geodata("bc-airports", crs = 3857)
bcdc_get_geodata("bc-airports", crs = 3857, query = "PHYSICAL_ADDRESS='Victoria, BC'")
bcdc_get_geodata("ground-water-wells", query = "OBSERVATION_WELL_NUMBER=108")
}
