% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcdc-web-services.R
\name{bcdc_get_geodata}
\alias{bcdc_get_geodata}
\title{Get data from the BC Web Feature Service (DEPRECATED)}
\usage{
bcdc_get_geodata(x = NULL, crs = 3005)
}
\arguments{
\item{x}{either a \code{bcdc_record} object (from the result of \code{bcdc_get_record()})
or a character string denoting the id of a resource (or the url).}

\item{crs}{the epsg code for the coordinate reference system. Defaults to \code{3005}
(B.C. Albers). See https://epsgi.io.}
}
\value{
an \code{sf} object
}
\description{
Get data from the BC Web Feature Service (DEPRECATED)
}
