% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcdc_search.R
\name{bcdc_get_record}
\alias{bcdc_get_record}
\title{Show a single B.C. Data Catalogue record}
\usage{
bcdc_get_record(id)
}
\arguments{
\item{id}{the id (name) of the record.}
}
\value{
A list containing the metadata for the record
}
\description{
Show a single B.C. Data Catalogue record
}
\examples{
bcdc_get_record("bc-airports")
}
