% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcdc_search.R
\name{bcdc_list_groups}
\alias{bcdc_list_groups}
\alias{bcdc_list_group_records}
\title{Retrieve group information for B.C. Data Catalogue}
\usage{
bcdc_list_groups()

bcdc_list_group_records(group)
}
\arguments{
\item{group}{Name of the group}
}
\description{
Returns a tibble of groups or records. Groups can be viewed here:
https://catalogue.data.gov.bc.ca/group or accessed directly from R using \code{bcdc_list_groups}
}
\section{Functions}{
\itemize{
\item \code{bcdc_list_groups()}: 

}}
\examples{
\donttest{
try(
  bcdc_list_group_records('environmental-reporting-bc')
)
}
}
