% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcdc_map.R
\name{bcdc_map}
\alias{bcdc_map}
\title{Get a map from a B.C. Data Catalogue record.}
\usage{
bcdc_map(id, epsg = 3005, query = NULL)
}
\arguments{
\item{id}{the name of the record}

\item{epsg}{the epsg code for the coordinate reference system. Default \code{3005}
(B.C. Albers). See https://epsgi.io.}

\item{query}{A valid \href{https://docs.geoserver.org/stable/en/user/tutorials/cql/cql_tutorial.html}{CQL or ECQL query}
to filter the results. Default \code{NULL} (return all objects)}
}
\value{
an \code{sf} map object
}
\description{
The data must be available as a wms service. See \code{bcdc_show(id)$resources})
}
\examples{
bcdc_map("bc-airports")
}
