% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcdc-web-services.R
\name{bcdc_preview}
\alias{bcdc_preview}
\title{Get map from the B.C. Web Mapping Service}
\usage{
bcdc_preview(record = NULL)
}
\arguments{
\item{record}{either a \code{bcdc_record} object (from the result of \code{bcdc_get_record()})
or a character string denoting the ID of a resource (or the url).

It is advised to use the permament ID for a record rather than the
human-readable name to guard against future name changes of the record.
If you use the human-readable name a warning will be issued once per
session. You can silence these warnings altogether by setting an option:
\code{options("silence_named_get_data_warning" = TRUE)} - which you can set
in your .Rprofile file so the option persists across sessions.}
}
\description{
Get map from the B.C. Web Mapping Service
}
\examples{
\dontrun{
bcdc_preview("regional-districts-legally-defined-administrative-areas-of-bc")
bcdc_preview("points-of-well-diversion-applications")
}
}
