% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcdc-web-services.R
\name{bcdc_query_geodata}
\alias{bcdc_query_geodata}
\title{Query data from the B.C. Web Feature Service}
\usage{
bcdc_query_geodata(record = NULL, crs = 3005)
}
\arguments{
\item{record}{either a \code{bcdc_record} object (from the result of \code{bcdc_get_record()})
or a character string denoting the name or id of a resource (or the url).

It is advised to use the permament id for a record rather than the
human-readable name to guard against future name changes of the record.
If you use the human-readable name a warning will be issued once per
session. You can silence these warnings altogether by setting an option:
\code{options("silence_named_get_data_warning" = TRUE)} - which you can set
in your .Rprofile file so the option persists across sessions.}

\item{crs}{the epsg code for the coordinate reference system. Defaults to \code{3005}
(B.C. Albers). See https://epsgi.io.}
}
\value{
an \code{sf} object
}
\description{
Queries features from the B.C. Web Feature Service. The data must be available as
a WMS/WFS service. See \code{bcdc_get_record(record)$resources}). Pulls features off the web.
See \code{bcdc_get_record(record)$resources}). If the record is greater than 10000 rows,
the response will be paginated. If you are querying layers of this size, expect
that the request will take quite a while.
}
\examples{

\dontrun{
bcdc_query_geodata("bc-airports", crs = 3857)
bcdc_query_geodata("bc-airports", crs = 3857) \%>\%
  filter(PHYSICAL_ADDRESS == 'Victoria, BC') \%>\%
  collect()
bcdc_query_geodata("ground-water-wells") \%>\%
  filter(OBSERVATION_WELL_NUMBER == 108)

## A moderately large layer
bcdc_query_geodata("bc-environmental-monitoring-locations")
bcdc_query_geodata("bc-environmental-monitoring-locations") \%>\%
  filter(PERMIT_RELATIONSHIP == "DISCHARGE")


## A very large layer
bcdc_query_geodata("terrestrial-protected-areas-representation-by-biogeoclimatic-unit")
}

}
