% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcdc_search.R
\name{bcdc_search_facets}
\alias{bcdc_search_facets}
\title{Get the valid values for a facet (that you can use in \code{\link[=bcdc_search]{bcdc_search()}})}
\usage{
bcdc_search_facets(facet = c("licence_id", "download_audience", "type",
  "res_format", "sector", "organization"))
}
\arguments{
\item{facet}{the facet(s) for which to retrieve valid values. Can be one or
more of:
\code{"licence_id", "download_audience", "type", "res_format", "sector", "organization"}}
}
\value{
a data frame of values for the selected facet
}
\description{
Get the valid values for a facet (that you can use in \code{\link[=bcdc_search]{bcdc_search()}})
}
\examples{
\dontrun{
bcdc_search_facets("type")
}
}
