% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcdc_search.R
\name{bcdc_show}
\alias{bcdc_show}
\title{Show the record for a single B.C. Data Catalogue package}
\usage{
bcdc_show(id)
}
\arguments{
\item{id}{the id (name) of the package.}
}
\value{
A list containing the metadata for the package
}
\description{
Show the record for a single B.C. Data Catalogue package
}
\examples{
bcdc_show("bc-airports")
}
