% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcdc_search.R
\name{bcdc_tidy_resources}
\alias{bcdc_tidy_resources}
\title{Provide metadata for all the resources from a single B.C. Data Catalogue record}
\usage{
bcdc_tidy_resources(x)
}
\arguments{
\item{x}{a list containing the metadata for a record
(result of running bcdc_get_record() function)}
}
\value{
A data frame containing the metadata for all the resources for a record
}
\description{
Provide metadata for all the resources from a single B.C. Data Catalogue record
}
\examples{
\donttest{
airports <- bcdc_get_record("bc-airports")
bcdc_tidy_resources(airports)
}
}
