% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-as_tibble.R, R/utils-classes.R,
%   R/utils-collect.R
\name{as_tibble}
\alias{as_tibble}
\alias{collect.bcdc_promise}
\alias{as_tibble.bcdc_promise}
\alias{collect}
\title{as_tibble}
\usage{
\method{collect}{bcdc_promise}(x, ...)

\method{as_tibble}{bcdc_promise}(x, ...)
}
\arguments{
\item{x}{object of class \code{bcdc_promise}}
}
\description{
See \code{tibble::\link[tibble]{as_tibble}} for details.

After tuning a query, \code{collect()} is used to actually bring the data into memory.
This will retrieve an sf object into R. The \code{as_tibble()} function can be used
interchangeably with \code{collect} which matches \code{dbplyr} behaviour.

See \code{dplyr::\link[dplyr:compute]{collect}} for details.
}
\examples{
\donttest{
try(
  bcdc_query_geodata("bc-airports") \%>\%
    collect()
)

try(
  bcdc_query_geodata("bc-airports") \%>\%
    as_tibble()
)
}

}
\keyword{internal}
