% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-classes.R, R/utils-collect.R
\name{collect.bcdc_promise}
\alias{collect.bcdc_promise}
\alias{collect}
\title{Force collection of Web Service request from B.C. Data Catalogue}
\usage{
\method{collect}{bcdc_promise}(x, ...)
}
\arguments{
\item{x}{object of class bcdc_promise}
}
\description{
After tuning a query, \code{collect()} is used to actually bring the data into memory.
This will retrieve an sf object into R.

See \code{dplyr::\link[dplyr]{collect}} for details.
}
\section{Methods (by class)}{
\itemize{
\item \code{bcdc_promise}: collect.bcdc_promise
}}

\examples{
\donttest{
bcdc_query_geodata("bc-airports") \%>\%
  collect()
}

}
\keyword{internal}
