% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-classes.R, R/utils-collect.R
\name{collect.bcdc_promise}
\alias{collect.bcdc_promise}
\alias{collect}
\title{Force collection of WFS request from BC Data Catalogue}
\usage{
\method{collect}{bcdc_promise}(x, ...)
}
\arguments{
\item{x}{object of class bcdc_promise}
}
\description{
Retrieve an sf object from data catalogue.

See \code{dplyr::\link[dplyr]{collect}} for details.
}
\section{Methods (by class)}{
\itemize{
\item \code{bcdc_promise}: collect.bcdc_promise
}}

\keyword{internal}
