% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cql-translator.R
\name{CQL}
\alias{CQL}
\title{CQL escaping}
\usage{
CQL(...)
}
\arguments{
\item{...}{Character vectors that will be comined into a single CQL statement.}
}
\value{
An object of class \code{c("CQL", "SQL")}
}
\description{
Write a CQL expression to escape its inputs, and return a CQL/SQL object.
Used when writing filter expressions in \code{\link[=bcdc_query_geodata]{bcdc_query_geodata()}}.
}
\details{
See \href{https://docs.geoserver.org/stable/en/user/tutorials/cql/cql_tutorial.html}{the CQL/ECQL for Geoserver website}.
}
\examples{
CQL("FOO > 12 & NAME LIKE 'A&'")
}
