% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-classes.R, R/utils-show-query.R
\name{show_query.bcdc_promise}
\alias{show_query.bcdc_promise}
\alias{show_query}
\title{Show SQL used for WFS request from BC Data Catalogue}
\usage{
\method{show_query}{bcdc_promise}(x, ...)
}
\arguments{
\item{x}{object of class bcdc_promise}
}
\description{
Display WFS query SQL

See \code{dplyr::\link[dplyr]{show_query}} for details.
}
\section{Methods (by class)}{
\itemize{
\item \code{bcdc_promise}: show_query.bcdc_promise
}}

\keyword{internal}
