% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cql-translator.R
\docType{class}
\name{wfsConnection-class}
\alias{wfsConnection-class}
\alias{dbQuoteIdentifier,wfsConnection,character-method}
\alias{dbQuoteIdentifier,wfsConnection,SQL-method}
\alias{dbQuoteString,wfsConnection,character-method}
\alias{dbQuoteString,wfsConnection,SQL-method}
\title{wfsConnection class}
\usage{
\S4method{dbQuoteIdentifier}{wfsConnection,character}(conn, x)

\S4method{dbQuoteIdentifier}{wfsConnection,SQL}(conn, x)

\S4method{dbQuoteString}{wfsConnection,character}(conn, x)

\S4method{dbQuoteString}{wfsConnection,SQL}(conn, x)
}
\description{
wfsConnection class
}
\keyword{internal}
